WIP send and receive encrypted payloads to HSBC
This work-in-progress adds a number of changes to the HSBC direct debit authorisation implementation:

- an `ActiveMerchant::Crypto` class that uses `gpgme` to encrypt and sign a payload prior to sending to HSBC, and decrypt and verify the ciphertext we get back from the response;
- small changes to the `add_direct_debit_authorisation_data` method that builds the payload;
- ensure that we always send the `application/json` content type header;

I've also redacted the remote test data, as it contains details that should probably not be in a public repo. The fixures changes that contain the development GPG public and private keys are also not committed.

The base docker container we use contains gpg and the gpgme required binaries to build the gem 🎉

Refs waysact/evergiving#6651