Add OTP confirmation method and small refactor
Add the ability to confirm the authorisation of the direct debit by calling the `authorisations/otp-confirmation` API endpoint with the data supplied via the SMS and API. I don't believe we have a test OTP password/mandate identification that we can use without getting a proper 2FA via SMS and entering it interactively, so I had to implement it with the `STDIN.gets` method in the remote test.

Did a bit of refactor to DRY the code that calls the ActiveMerchant::Crypto module (and fixed a comment in that one).

These changes have enabled me to generate a successful direct debit authorisation and confirm it, thus getting an SMS from the pre-prod system that states that HSBC received the direct debit authorisation setup request.

Refs waysact/evergiving#6651