Add support for EngagingNetworks
This is like JacksonRiver and other CRM/PaymentGateway concentrators. It's not technically a gateway, as it uses Vantiv/Worldpay in the background, combined with BBCRM.

The gateway supports creditcard and directdebit `purchase`, however given the comment above, you will see that the implementation contains support for recurring billing for each payment method, however those are handled by EN and not sent/expected differently. I've implemented this to match the approach we took with Convio, i.e. we expose only one method and the CRM does all the decision based on the data we pass.

EN supplies an `api_key` that we pass on to retrieve a session token during the authentication call, we then use that token in the headers of our subsequent interaction.

Refs waysact/evergiving#6789